Add extra recognized parameters for comment passwords
This commit adds functionality to processComment.php that allows
users to optionally provide a password with their comment, which
will be written to the database as an MD5 hash.

Users can then change their old comments, provided they set a
password, by providing the comment ID, the password, and a new
comment.